% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{apply_theme}
\alias{apply_theme}
\title{Apply a neat, customized ggplot2 theme}
\usage{
apply_theme(theme_type = "ipsum")
}
\arguments{
\item{theme_type}{Type of a base hrbrthemes theme to use. Currently only "ipsum" is supported.}
}
\description{
This function applies a customized ggplot2 theme (based on the hrbrthemes package)
to any ggplot graph in order to create neat and good looking visualizations.
}
