% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_decile_table}
\alias{calculate_decile_table}
\title{Calculate a decile breakdown}
\usage{
calculate_decile_table(df, binning, grouping, top_level = "1",
  n_bins = 10, format = FALSE, ...)
}
\arguments{
\item{df}{A data frame.}

\item{binning}{Variable for which binning should be applied.}

\item{grouping}{A two-level (binary) variable to calculate the ratio in each bin.}

\item{top_level}{Top level of the grouping variable. Defaults to "1".}

\item{n_bins}{Provide a number of bins. Defaults to 10.}

\item{format}{Should table printing be formatted with kable? Defaults to FALSE}

\item{...}{Additional grouping columns to calculate deciles}
}
\description{
This function calculates a decile table for any combination of numerical and categorical variables.
}
\examples{
library(tidyverse)

diamonds_filter <- diamonds \%>\% filter(cut \%in\% c("Ideal", "Premium"))

diamonds_filter \%>\%
  calculate_decile_table(price, cut, "Ideal")

diamonds_filter \%>\%
  calculate_decile_table(binning = price,
                         grouping = cut,
                         top_level = "Ideal",
                         n_bins = 10,
                         format = FALSE,
                         color)
}
