% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diamonds}
\alias{diamonds}
\title{Diamonds dataset used for examples}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 53940 rows and 10 columns.}
\usage{
diamonds
}
\description{
See \code{\link[ggplot2]{diamonds}} for details.
}
\keyword{datasets}
