% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{first_to_lower}
\alias{first_to_lower}
\title{Convert strings first letter to lowercase}
\usage{
first_to_lower(x)
}
\arguments{
\item{x}{A data frame or vector.}
}
\description{
This function converts your data frame column names or character vector elements
first letter to lowercase. It can be used in a pipe.
}
\examples{
first_to_lower(c("first", "second"))
}
