% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{first_to_upper}
\alias{first_to_upper}
\title{Convert strings first letter to uppercase}
\usage{
first_to_upper(x)
}
\arguments{
\item{x}{A data frame or vector.}
}
\description{
This function converts your data frame column names or character vector elements
first letter to uppercase. It can be used in a pipe.
}
\examples{
first_to_upper(c("first", "second"))

diamonds \%>\%
   first_to_upper()
}
