% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{number_ticks}
\alias{number_ticks}
\title{Specify equal sized axis ticks}
\usage{
number_ticks(n = 10)
}
\arguments{
\item{n}{Number of desired ticks on an axis. Defaults to 10.}
}
\description{
This function creates equally spaced axis ticks for ggplot graphs. Should be used as input
for the "break" argument of scale_continuous function in a ggplot function.
}
