% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bars.R
\name{plot_bars}
\alias{plot_bars}
\title{Plot bars of numerical variables}
\usage{
plot_bars(df, x, y = NULL, y_prop = FALSE, x_type = "num",
  fill = NULL, facet = NULL, binwidth = NULL, position = "stack",
  stat = "count", fct_order = FALSE, angle = 0, title = TRUE,
  subtitle = NULL, caption = NULL, lab_x = "Value range",
  lab_y = "Proportion", legend = TRUE, vline = c(NaN), alpha = 0.7,
  quantile_low = 0, quantile_high = 1, theme_type = "ipsum",
  palette = "awtools")
}
\arguments{
\item{df}{A data frame.}

\item{x}{A numeric/ categorical variable to plot the bar plot.}

\item{y}{A numeric variable which contains summarised y values (before plotting). Use only with stat = "identity".}

\item{y_prop}{A logical variable to choose between counts/ proportion on the y axis. Defaults to FALSE (proportion).}

\item{x_type}{Character identifier for type of the variable x defined above: "num" for numeric (plots histogram) and "char" for character (plots bar chart). Defauls to "num".}

\item{fill}{Select an additional grouping variable to be used for plotting. Defaults to NULL.}

\item{facet}{Select an additional faceting variable to create facets. Defaults to NULL.}

\item{binwidth}{Select binwidth, defaults to NULL and lets ggplot select the optimal binwidth.}

\item{position}{Select the position of the barplot from: "stack" (default), "dodge" or "fill".}

\item{stat}{Character identifier for whether the data is already grouped ("identity") or if the function needs to aggregate data at the level of x ("count").}

\item{fct_order}{Should the factors be reordered by their frequency? Defaults to FALSE.}

\item{angle}{Select the rotation angle for the x axis labels. Defaults to 0.}

\item{title}{Should the plot title appear automatically. Defaults to TRUE.}

\item{subtitle}{Text that is displayed on the subtitle. Defaults to NULL.}

\item{caption}{Text that is displayed on the caption. Defaults to NULL.}

\item{lab_x}{Text that is displayed on the x axis. Defaults to "Level".}

\item{lab_y}{Text that is displayed on the y axis. Defaults to "Value range".}

\item{legend}{Should the plot legend appear automatically. Defaults to TRUE.}

\item{vline}{Should any horizontal lines be added to the plot. Defaults to c(NaN).}

\item{alpha}{Select plot fill transparency. Defaults to 0.7.}

\item{quantile_low}{Select lower percentile for outliers exclusion. Defaults to 0.0\%.}

\item{quantile_high}{Select upper percentile for outliers exclusion. Defaults to 1.0\%.}

\item{theme_type}{Select a theme type from themes available in the \code{\link{apply_theme}} function. Defaults to "ipsum".}

\item{palette}{Select a palette type with the \code{\link{select_palette}} function. Defaults to "awtools".}
}
\description{
This function creates nicely formatted, standardized bar-plots.
}
\examples{
library(tidyverse)

diamonds_sum <- diamonds \%>\%
  group_by(cut) \%>\%
  summarise(mean_carat = mean(carat, na.rm = TRUE))

}
