% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Plot box-plots of numerical variables}
\usage{
plot_boxplot(df, x, y, fill = NULL, facet = NULL, ticks = 10,
  angle = 0, title = TRUE, subtitle = NULL, caption = NULL,
  lab_x = "Level", lab_y = "Value range", legend = TRUE,
  vline = c(NaN), alpha = 0.7, quantile_low = 0, quantile_high = 1,
  theme_type = "ipsum", palette = "awtools")
}
\arguments{
\item{df}{A data frame.}

\item{x}{A categorical variable for the x axis groupings.}

\item{y}{A numerical variable for the y axis levels.}

\item{fill}{Select an additional grouping variable to be used for plotting. Defaults to NULL.}

\item{facet}{Select an additional faceting variable to create facets. Defaults to NULL.}

\item{ticks}{Select the number of ticks on the x and y axis with the \code{\link{number_ticks}} function. Defaults to 10.}

\item{angle}{Select the rotation angle for the x axis labels. Defaults to 0.}

\item{title}{Should the plot title appear automatically. Defaults to TRUE.}

\item{subtitle}{Text that is displayed on the subtitle. Defaults to NULL.}

\item{caption}{Text that is displayed on the caption. Defaults to NULL.}

\item{lab_x}{Text that is displayed on the x axis. Defaults to "Level".}

\item{lab_y}{Text that is displayed on the y axis. Defaults to "Value range".}

\item{legend}{Should the plot legend appear automatically. Defaults to TRUE.}

\item{vline}{Should any horizontal lines be added to the plot. Defaults to c(NaN).}

\item{alpha}{Select plot fill transparency. Defaults to 0.7.}

\item{quantile_low}{Select lower percentile for outliers exclusion. Defaults to 0.0\%.}

\item{quantile_high}{Select upper percentile for outliers exclusion. Defaults to 1.0\%.}

\item{theme_type}{Select a theme type from themes available in the \code{\link{apply_theme}} function. Defaults to "ipsum".}

\item{palette}{Select a palette type with the \code{\link{select_palette}} function. Defaults to "awtools".}
}
\description{
This function creates nicely formatted, standardized box-plots.
}
\examples{
diamonds \%>\%
  plot_boxplot(x = cut,
               y = carat)

diamonds \%>\%
  plot_boxplot(x = cut,
               y = carat,
               fill = color)

diamonds \%>\%
  plot_boxplot(x = cut,
               y = carat,
               fill = cut,
               facet = color)

diamonds \%>\%
  plot_boxplot(x = cut,
               y = carat,
               fill = cut,
               title = "Write your title here",
               subtitle = "Write your subtitle here",
               caption = "Write your caption here",
               lab_x = "Carat",
               alpha = .5,
               vline = 1)
}
