% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Plot a correlation matrix of numerical variables}
\usage{
plot_correlation(df, method = "spearman", order = "alphabet",
  label_size = 0.7, number_size = 0.9)
}
\arguments{
\item{df}{A data frame.}

\item{method}{A character string indicating which correlation coefficient (or covariance) should be computed. Options are: "spearman" (default), "pearson" or "kendall".}

\item{order}{Ordering method of the correlation matrix. Options are: "alphabet" (default) and "hclust".}

\item{label_size}{Size of the text label. Defaults to 0.7.}

\item{number_size}{Size of the correlation number. Defaults to 0.9.}
}
\description{
This function creates a nicely formatted, standardised correlation matrix of numerical variables.
Make sure that long variables names are are shortened before using the function for easier interpretation.
}
\examples{
diamonds \%>\%
  plot_correlation()

}
