% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deciles.R
\name{plot_deciles}
\alias{plot_deciles}
\title{Plot decile plots of numerical variables}
\usage{
plot_deciles(df, x = decile, y = ratio, facet = NULL, ticks = 10,
  angle = 0, title = TRUE, subtitle = NULL, caption = NULL,
  lab_x = "Decile", lab_y = "Ratio", legend = TRUE, alpha = 0.7,
  quantile_low = 0, quantile_high = 1, theme_type = "ipsum",
  palette = "awtools")
}
\arguments{
\item{df}{A data frame.}

\item{x}{A categorical variable for the x axis groupings. Defaults to 'decile'.}

\item{y}{A numerical variable for the y axis levels. Defaults to 'ratio'.}

\item{facet}{Select an additional faceting variable to create facets. Defaults to NULL.}

\item{ticks}{Select the number of ticks on the x and y axis with the \code{\link{number_ticks}} function. Defaults to 10.}

\item{angle}{Select the rotation angle for the x axis labels. Defaults to 0.}

\item{title}{Should the plot title appear automatically. Defaults to TRUE.}

\item{subtitle}{Text that is displayed on the subtitle. Defaults to NULL.}

\item{caption}{Text that is displayed on the caption. Defaults to NULL.}

\item{lab_x}{Text that is displayed on the x axis. Defaults to "Decile".}

\item{lab_y}{Text that is displayed on the y axis. Defaults to "Percentage".}

\item{legend}{Should the plot legend appear automatically. Defaults to TRUE.}

\item{alpha}{Select plot fill transparency. Defaults to 0.7.}

\item{quantile_low}{Select lower percentile for outliers exclusion. Defaults to 0.0\%.}

\item{quantile_high}{Select upper percentile for outliers exclusion. Defaults to 1.0\%.}

\item{theme_type}{Select a theme type from themes available in the \code{\link{apply_theme}} function. Defaults to "ipsum".}

\item{palette}{Select a palette type with the \code{\link{select_palette}} function. Defaults to "awtools".}
}
\description{
This function creates nicely formatted, standardised decile plots. Prior to calling the function
the data should only be in a form of a decile table \code{\link{calculate_decile_table}}. Above
each bar the median value of the bin is displayed.
}
\examples{
library(tidyverse)

diamonds_filter <- diamonds \%>\% filter(cut \%in\% c("Ideal", "Premium"))

diamonds_filter \%>\%
  calculate_decile_table(price, cut, "Ideal") \%>\%
  plot_deciles()

diamonds_filter \%>\%
  calculate_decile_table(binning = price,
                         grouping = cut,
                         top_level = "Ideal",
                         n_bins = 10,
                         format = FALSE,
                         color) \%>\%
  plot_deciles(facet = color)
}
