% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot density of numerical variables}
\usage{
plot_density(df, x, fill = NULL, facet = NULL, position = "identity",
  ticks = 5, angle = 0, title = TRUE, subtitle = NULL,
  caption = NULL, lab_x = "Value range", lab_y = "Density",
  legend = TRUE, vline = c(NaN), alpha = 0.7, quantile_low = 0,
  quantile_high = 1, theme_type = "ipsum", palette = "awtools")
}
\arguments{
\item{df}{A data frame.}

\item{x}{A numerical variable to plot its density.}

\item{fill}{Select an additional grouping variable to be used for density plotting. Defaults to NULL.}

\item{facet}{Select an additional faceting variable to create facets. Defaults to NULL.}

\item{position}{Select the positional adjustment of the plot.  Possible values are: "identity" (default), "stack" or "fill"}

\item{ticks}{Select the number of ticks on the x and y axis with the \code{\link{number_ticks}} function. Defaults to 10.}

\item{angle}{Select the rotation angle for the x axis labels. Defaults to 0.}

\item{title}{Should the plot title appear automatically. Defaults to TRUE.}

\item{subtitle}{Text that is displayed on the subtitle. Defaults to NULL.}

\item{caption}{Text that is displayed on the caption. Defaults to NULL.}

\item{lab_x}{Text that is displayed on the x axis. Defaults to "Value range".}

\item{lab_y}{Text that is displayed on the y axis. Defaults to "Density".}

\item{legend}{Should the plot legend appear automatically. Defaults to TRUE.}

\item{vline}{Should any vertical lines be added to the plot. Defaults to c(NaN).}

\item{alpha}{Select plot fill transparency. Defaults to 0.7.}

\item{quantile_low}{Select lower percentile for outliers exclusion. Defaults to 0.0\%.}

\item{quantile_high}{Select upper percentile for outliers exclusion. Defaults to 1.0\%.}

\item{theme_type}{Select a theme type from themes available in the \code{\link{apply_theme}} function. Defaults to "ipsum".}

\item{palette}{Select a palette type with the \code{\link{select_palette}} function. Defaults to "awtools".}
}
\description{
This function creates nicely formatted, standardized density plots.
}
\examples{
diamonds \%>\%
  plot_density(x = carat)

diamonds \%>\%
  plot_density(x = carat,
               fill = cut)

diamonds \%>\%
  plot_density(x = carat,
               fill = cut,
               position = "stack")

diamonds \%>\%
  plot_density(x = carat,
               fill = cut,
               position = "fill")

diamonds \%>\%
  plot_density(x = carat,
               facet = cut)

diamonds \%>\%
  plot_density(x = carat,
               fill = cut,
               facet = cut,
               alpha = .5)

diamonds \%>\%
  plot_density(x = carat,
               fill = cut,
               facet = cut,
               title = "Write your title here",
               subtitle = "Write your subtitle here",
               caption = "Write your caption here",
               lab_x = "Carat",
               alpha = .5,
               vline = 1)
}
