% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lines.R
\name{plot_line}
\alias{plot_line}
\title{Plot lines of numerical variables. Usefull especially for time-series data}
\usage{
plot_line(df, x, y, fill = NULL, facet = NULL, ticks = 10,
  angle = 0, title = TRUE, subtitle = NULL, caption = NULL,
  lab_x = "Value range", lab_y = "Value range", legend = TRUE,
  hline = c(NaN), alpha = 0.7, limit_min = NA, limit_max = NA,
  theme_type = "ipsum", palette = "awtools")
}
\arguments{
\item{df}{A data frame.}

\item{x}{A categorical variable for the x axis groupings.}

\item{y}{A numerical variable for the y axis levels.}

\item{fill}{Select an additional grouping variable to be used for plotting. Defaults to NULL.}

\item{facet}{Select an additional faceting variable to create facets. Defaults to NULL.}

\item{ticks}{Select the number of ticks on the x and y axis with the \code{\link{number_ticks}} function. Defaults to 10.}

\item{angle}{Select the rotation angle for the x axis labels. Defaults to 0.}

\item{title}{Should the plot title appear automatically. Defaults to TRUE.}

\item{subtitle}{Text that is displayed on the subtitle. Defaults to NULL.}

\item{caption}{Text that is displayed on the caption. Defaults to NULL.}

\item{lab_x}{Text that is displayed on the x axis. Defaults to "Value range".}

\item{lab_y}{Text that is displayed on the y axis. Defaults to "Value range".}

\item{legend}{Should the plot legend appear automatically. Defaults to TRUE.}

\item{hline}{Should any horizontal lines be added to the plot. Defaults to c(NaN).}

\item{alpha}{Select plot fill transparency. Defaults to 0.7.}

\item{limit_min}{Select lower limit for the y scale. Defaults to NA.}

\item{limit_max}{Select upper limit for the y scale. Defaults to NA.}

\item{theme_type}{Select a theme type from themes available in the \code{\link{apply_theme}} function. Defaults to "ipsum".}

\item{palette}{Select a palette type with the \code{\link{select_palette}} function. Defaults to "awtools".}
}
\description{
This function creates nicely formatted, standardised line plots. Color and group parameters for geom_line and
geom_point are automatically inherited from the fill parameter.
}
\examples{
library(tidyverse)

tibble(
  time = 1:20,
  value = rnorm(20, 0.5, 2)
  ) \%>\%
  plot_line(
    x = time,
    y = value,
    ticks = 10,
    hline = 0.05
  )

tibble(
  time = 1:20,
  value = rnorm(20, 0.5, 2)
  ) \%>\%
  plot_line(
    x = time,
    y = value,
    ticks = 10,
    hline = 0.05,
    limit_min = -2,
    limit_max = 2
  )
}
