% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{select_palette}
\alias{select_palette}
\title{Apply a neat, appealing pallete scheme}
\usage{
select_palette(palette = "awtools")
}
\arguments{
\item{palette}{Select a palette. Available options for discrete palettes are: "awtools" (8 discrete colors)
and for continuous paletter: "berlin" or "lajolla" (60 continuous colors each). Defaults to "awtools".}
}
\description{
Palettes are based on the list of available color schemes: \code{\link{https://github.com/EmilHvitfeldt/r-color-palettes}}.
}
